+++ src/main/kotlin/cz/machovec/lekovyportal/core/domain/distribution/DistFromMahs.kt
@@ -12,6 +12,7 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 
 @Entity
@@ -43,5 +44,5 @@ data class DistFromMahs(
     val movementType: MovementType,
 
     @Column(name = "package_count", nullable = false)
-    val packageCount: Int
+    val packageCount: BigDecimal
 )
