+++ src/main/kotlin/cz/machovec/lekovyportal/core/dto/distribution/MonthlyDistributorProductMovementCountDto.kt
@@ -2,6 +2,7 @@ package cz.machovec.lekovyportal.core.dto.distribution
 
 import cz.machovec.lekovyportal.core.domain.distribution.DistributorPurchaserType
 import cz.machovec.lekovyportal.core.domain.distribution.MovementType
+import java.math.BigDecimal
 
 data class MonthlyDistributorProductMovementCountDto(
     val year: Int,
@@ -9,5 +10,5 @@ data class MonthlyDistributorProductMovementCountDto(
     val medicinalProductId: Long,
     val purchaserType: DistributorPurchaserType,
     val movementType: MovementType,
-    val packageCount: Long
+    val packageCount: BigDecimal
 )
